bk_eenheid is used instead of bk_bron; 22 points automatically set to wpeil and 0's; points are sorted. Tests are adapted.
Bk_eenheid is a better indicator of the unit in which the data is measured.
Points are sorted by profile id first, and distance next.
Tests are adapted to test a situation that occurs often, in which for 1 and 2 points, ok_eenheid and ok_bron are filled (they shouldn't be). As a result, bk values are not copied properly. Now it is working correctly.