Fix code, edit tests and make comments
Fix code to correctly calculate distances when there is only one profile per line.
Sort the points, based on the distance from the start of the line; base the rest of the code on this sorted points collection.
Add tests on total length.
Comment code.

Code works correctly in tests and also when tested in ArcGIS, this closes issue TijhuisArcGisTools/#31